Reorganize varscore/ into functional subpackages
Flat layout was getting crowded and "scoring" is about to become plural
(AlphaMissense, ChromBPNet, SpliceAI). Group the package by function, with each
scorer as a self-contained subpackage under scoring/<x>/.

Layout (all moves via git mv; history preserved):
- core/        io.py (← utils/io_utils), logging.py (← utils/logging_config)
- preprocessing/ validate, region_filter, pipeline (← validate_variants,
                 variant_region_filter, variant_preprocessing)
- annotation/  annotate (← annotations), regions (← utils/region_utils),
               maf (← utils/variant_utils), conservation (← utils/conservation_utils)
- scoring/alphamissense/ lookup (← utils/alphamissense_utils), score (← alphamissense_scoring)
- scoring/chrombpnet/    model (← utils/chrombpnet_utils), score (← scoring),
                         predictions (← model_predictions), ingest (← ingest_model),
                         interpret/ (← shap/*)
- prioritization.py stays at root; scripts/ unchanged.

utils/ and shap/ are dissolved into the groups they served. Internal imports use
aliases (import ... as region_utils), so only import lines changed, not call
sites. Organizing scorers by domain (not lookup-vs-model mechanism) keeps a
future scorer's two implementations + shared output-column contract co-located.

CLI module paths change accordingly (e.g. `python -m varscore.scoring` ->
`python -m varscore.scoring.chrombpnet.score`); all in-repo callers updated
(Dockerfile, docs, CI, tests). No back-compat shims — clean break, pre-1.0.
Version bumped to 0.0.2a0.

Validated: wheel ships all 9 subpackages; clean core install is TF-free; all 39
tests pass against the installed wheel.
