+++ varscore/__init__.py
@@ -1,9 +1,9 @@
 """varscore: common functions for variant effect prediction & prioritization.
 
 The names re-exported here are the TensorFlow-free core API (annotation,
-region classification, lookups, prioritization). Model scoring / SHAP live in
-``varscore.scoring``, ``varscore.model_predictions`` and ``varscore.shap`` and
-require the ``[model]`` extra; import those modules directly.
+region classification, lookups, prioritization). Model scoring / SHAP live
+under ``varscore.scoring.chrombpnet`` (``.score``, ``.predictions``,
+``.interpret``) and require the ``[model]`` extra; import those modules directly.
 """
 
 from importlib.metadata import PackageNotFoundError, version
@@ -13,7 +13,7 @@ try:
 except PackageNotFoundError:  # running from a source tree without an install
     __version__ = "0.0.0+dev"
 
-from varscore.annotations import (
+from varscore.annotation.annotate import (
     AnnotatedVariant,
     VariantAnnotationInput,
     annotate_variant,
