+++ varscore/scoring/chrombpnet/interpret/prepare_variant_plotting.py
@@ -8,7 +8,7 @@ import io
 import multiprocessing
 import os
 
-import varscore.utils.io_utils as io_utils
+import varscore.core.io as io_utils
 
 
 #################
